+++ lib/specjour/connection.rb
@@ -3,10 +3,10 @@ module Specjour
     include Protocol
     extend Forwardable
 
-    attr_reader :uri
+    attr_reader :uri, :retries
     attr_writer :socket
 
-    def_delegators :socket, :flush, :closed?, :gets, :each
+    def_delegators :socket, :flush, :close, :closed?, :gets, :each
 
     def self.wrap(established_connection)
       host, port = established_connection.peeraddr.values_at(3,1)
@@ -17,6 +17,7 @@ module Specjour
 
     def initialize(uri)
       @uri = uri
+      @retries = 0
     end
 
     alias to_str to_s
@@ -26,17 +27,11 @@ module Specjour
     end
 
     def disconnect
-      socket.close
+      socket.close if socket && !socket.closed?
     end
 
     def socket
       @socket ||= connect
-    end
-
-    def timeout(&block)
-      Timeout.timeout(2, &block)
-    rescue Timeout::Error
-      raise Error, "Connection to dispatcher timed out", []
     end
 
     def next_test
@@ -53,12 +48,16 @@ module Specjour
     end
 
     def puts(arg='')
-      print(arg << "\n")
+      will_reconnect do
+        print(arg << "\n")
+      end
     end
 
     def send_message(method_name, *args)
-      print([method_name, *args])
-      flush
+      will_reconnect do
+        print([method_name, *args])
+        flush
+      end
     end
 
     protected
@@ -66,8 +65,6 @@ module Specjour
     def connect_socket
       @socket = TCPSocket.open(uri.host, uri.port)
     rescue Errno::ECONNREFUSED => error
-      Specjour.logger.debug "Could not connect to #{uri.to_s}\n#{error.inspect}"
-      sleep 0.1
       retry
     end
 
@@ -76,12 +73,18 @@ module Specjour
       connect
     end
 
+    def timeout(&block)
+      Timeout.timeout(1.0, &block)
+    rescue Timeout::Error
+    end
+
     def will_reconnect(&block)
       block.call
-    rescue SystemCallError => error
+    rescue SystemCallError, IOError => error
       unless Specjour.interrupted?
+        @retries += 1
         reconnect
-        retry
+        retry if retries <= 5
       end
     end
   end