+++ lib/specjour/rspec.rb
@@ -11,11 +11,6 @@ module Specjour
     require 'specjour/rspec/shared_example_group_ext'
 
     ::RSpec::Core::Runner.disable_autorun!
-
-    def self.wants_to_quit
-      if defined?(::RSpec) && ::RSpec.respond_to?(:wants_to_quit=)
-        ::RSpec.wants_to_quit = true
-      end
-    end
+    ::RSpec::Core::Runner.class_eval "def self.trap_interrupt;end"
   end
 end