+++ lib/specjour/worker.rb
@@ -1,24 +1,17 @@
 module Specjour
-  require 'specjour/rspec'
-  require 'specjour/cucumber'
 
   class Worker
     include Protocol
     include SocketHelper
     attr_accessor :printer_uri
-    attr_reader :project_path, :number, :preload_spec, :preload_feature, :task
+    attr_reader :number
 
     def initialize(options = {})
       ARGV.replace []
       $stdout = StringIO.new if options[:quiet]
-      @project_path = options[:project_path]
       @number = options[:number].to_i
-      @preload_spec = options[:preload_spec]
-      @preload_feature = options[:preload_feature]
-      @task = options[:task]
       self.printer_uri = options[:printer_uri]
       set_env_variables
-      Dir.chdir(project_path)
       Specjour.load_custom_hooks
     end
 
@@ -27,13 +20,10 @@ module Specjour
     end
 
     def prepare
-      load_app
       Configuration.prepare.call
-      Kernel.exit!
     end
 
     def run_tests
-      load_app
       Configuration.after_fork.call
       run_times = Hash.new(0)
 
@@ -43,15 +33,12 @@ module Specjour
         time = Benchmark.realtime { run_test test }
         profile(test, time)
         run_times[test_type(test)] += time
+        connection.send_message(:done)
       end
 
       send_run_times(run_times)
-      connection.send_message(:done)
+    ensure
       connection.disconnect
-    end
-
-    def start
-      send task
     end
 
     protected
@@ -60,22 +47,13 @@ module Specjour
       @connection ||= printer_connection
     end
 
-    def load_app
-      RSpec::Preloader.load(preload_spec) if preload_spec
-      Cucumber::Preloader.load(preload_feature) if preload_feature
-    rescue StandardError => exception
-      $stderr.puts "Caught exception: #{exception.class} #{exception.message}"
-      Specjour.logger.debug exception.backtrace.join("\n")
-      $stderr.puts "Proceeding... you may need to re-run the dispatcher."
-    end
-
     def printer_connection
       Connection.new printer_uri
     end
 
     def print_status(test)
       status = "[#{ENV['TEST_ENV_NUMBER']}] Running #{test}"
-      Specjour.logger.debug status
+      $stdout.puts status
       $PROGRAM_NAME = "specjour#{status}"
     end
 
@@ -97,11 +75,11 @@ module Specjour
     end
 
     def run_feature(feature)
-      Specjour::Cucumber::Runner.run(feature, connection)
+      Cucumber::Runner.run(feature, connection)
     end
 
     def run_spec(spec)
-      Specjour::RSpec::Runner.run(spec, connection)
+      RSpec::Runner.run(spec, connection)
     end
 
     def send_run_times(run_times)